+++ src/main/java/com/ipiecoles/java/java350/service/EmployeService.java
@@ -19,7 +19,8 @@ public class EmployeService {
 
     @Autowired
     private EmployeRepository employeRepository;
-    
+
+    //private Logger logger = LoggerFactory.getLogger(this.getClass());
     private static final Logger LOG = LoggerFactory.getLogger(EmployeService.class);
