+++ src/test/java/com/ipiecoles/java/java350/repository/EmployeRepositoryIntegrationTest.java
@@ -1,0 +1,44 @@
+package com.ipiecoles.java.java350.repository;
+
+import com.ipiecoles.java.java350.exception.EmployeException;
+import com.ipiecoles.java.java350.model.Employe;
+import com.ipiecoles.java.java350.model.Entreprise;
+import com.ipiecoles.java.java350.service.EmployeService;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.time.LocalDate;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest
+public class EmployeRepositoryIntegrationTest {
+
+    @Autowired
+    EmployeService employeService;
+
+    @Autowired
+    private EmployeRepository employeRepository;
+
+    @BeforeEach
+    @AfterEach
+    public void setup(){
+        employeRepository.deleteAll();
+    }
+
+    @Test
+    public void integrationAvgPerformanceWhereMatriculeStartsWit() throws EmployeException {
+        //Given
+        employeRepository.save(new Employe("Doe", "John", "C12345", LocalDate.now(), Entreprise.SALAIRE_BASE, 1, 1.0));
+        employeRepository.save(new Employe("Doe", "John", "C23456", LocalDate.now(), Entreprise.SALAIRE_BASE, 2, 1.0));
+        employeRepository.save(new Employe("Doe", "John", "C34567", LocalDate.now(), Entreprise.SALAIRE_BASE, 3, 1.0));
+
+        //When/Then
+        Assertions.assertEquals(2, employeRepository.avgPerformanceWhereMatriculeStartsWith("C").intValue());
+    }
+}
